Improved exception handling for parallel::sort
